04302025 - PO Return Controller: added intvalue for get receive qty
- added intvalue for get receive qty

+++ src/main/java/com/rmj/guanzongroup/sidebarmenus/controller/PurchaseOrderReturn_ConfirmationController.java
@@ -546,7 +546,7 @@ public class PurchaseOrderReturn_ConfirmationController implements Initializable
                     }
                     if (poPurchaseReturnController.Detail(pnDetail).getQuantity() != null
                             && !"".equals(poPurchaseReturnController.Detail(pnDetail).getQuantity())) {
-                        if (poPurchaseReturnController.getReceiveQty(pnDetail) < Integer.valueOf(lsValue)) {
+                        if (poPurchaseReturnController.getReceiveQty(pnDetail).intValue() < Integer.valueOf(lsValue)) {
                             ShowMessageFX.Warning(null, pxeModuleName, "Return quantity cannot be greater than the order quantity.");
                             poPurchaseReturnController.Detail(pnDetail).setQuantity(0);
                             tfReceiveQuantity.requestFocus();
